Initial code refactor
    - rename main file from Datto_backup_check.py to main.py
    - move configs to a new file config.py
    - move Datto class to new module
    - add 'if __name__ == '__main__:' to start things
    - add .gitignore
