+++ .gitignore
@@ -1,0 +1,2 @@
+__pycache__
+venv
